Moved HexStringToBytes into a SixLabors.ImageSharp.Common.Helpers.HexConverter.
